Qt.core: Make moc rule catch changes to moc_*.cpp inclusions
The problem was that the moc_xxx.cpp artifact gets created from the
xxx.h input, but its file tags depend on whether it gets included by
some cpp file or not. Because changes to such cpp files do not affect
the xxx.h file in any way, the rule was not re-applied in the case such
includes got added or removed.
We fix this by making cpp files an auxiliary input to the moc rule, so
it gets re-applied on changes to those files.

Change-Id: I29b04fac3870f5a370cb7d894bc6ea56a63f6b0e
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>