Fix OOM crash in QRC scanner
We implicitely created a QByteArray with the file content without
passing the file size. This would result in copying the whole mmapped
file into memory, and potentially much more if there's no terminating
null byte.

Fix this by using QByteArray::fromRawData which we pass the correct file
size and which doesn't copy the data.

Task-number: QBS-1375
Change-Id: I35c4cceba64343550094c29298ff9b3617718dac
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>