Qt.core: Split up the moc rule
The possible results are quite different for running on headers vs
running on implementation files. This is now expressed by the different
rule interfaces. Also, we save some unnecessary re-applications of the
rule: In the case that a cpp file changes, only the variant of the rule
that takes header files needs to run again to scan for includes of the
respective moc_xxx.cpp file.

Change-Id: Iaef30a84dc26cfc796cfe7558c2e7744ec03a706
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>