RulesApplicator: Fix possible assertion failure
The RuleNode class disconnects removed input artifacts, which could
break an assertion in the RulesApplicator in the case that a rule that
can provide different tags to the same artifact switches back and forth
between these tags. Fix this by re-establishing the connection between
output artifact and parent rule node if necessary.

Change-Id: I8d50190cba782ffe63d3a73b3d5400d95c99ea0e
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>