Android: Enable building apps with native code and no multiplexing
Until now, we needed to either put the native part into its own product
or use multiplexing with the "APK product" serving as the aggregate. Now
it is also possible to use a single product without multiplexing, which
is a more natural approach in the case where there is only one
architecture.

Change-Id: I976168c99f75ad8e4940ac61f957c64ad29f5f5c
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>