qmake build: Opt out of qtquickcompiler functionality
The qtprofilesetup library contains a resource file with JavaScript
sources. Now if "CONFIG += qtquickcompiler" is set globally, as it
happens by default in Qt Creator, qmake believes that we want to compile
the js files, which leads to disaster.
Arguably, Qt should ignore the qtquickcompiler CONFIG value if the
product does not pull in QtQml, but until that happens, let's force-
disable the qtquickcompiler.

Fixes: QTCREATORBUG-21340
Change-Id: If4802277ab66b550b0c88e5678cadf59a32e68d2
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>