bare-metal: Add IAR EWARM toolchain support on Windows
This commit adds a basic support of the IAR Embedded Workbench
toolchain for the ARM processors on Windows host.

To use it with Qt Creator, it is enough to add there a desired Kit
with a custom IAR C/C++ compiler, and then set the following in the
Kit's Qbs profile settings:

* Key: qbs.toolchainType
* Value: iar

To specify the target CPU/FPU you need to set the cpp.driverFlags
property, like this:

    cpp.driverFlags: [
        "--cpu", "Cortex-M4",
        "--fpu", "VFPv4_sp"
    ]

Then these flags will be automatically passed to both compiler
and assembler.

To specify the linker flags you need to set the cpp.driverLinkerFlags
property instead of cpp.linkerFlags property, like this:

    cpp.driverLinkerFlags: ["--vfe"]

To add the linker script files you need to set the 'linkerscript'
tag, e.g. in the following way:

    Group {
        name: "Linker scripts"
        fileTags: ["linkerscript"]
        files: ["stm32f407xx_flash.icf"]
    }

Other properties can be used as usual, according to the EWARM
compiler documentation.

Tested with EWARM v8.20.2, v6.50.3 on Windows using:

* STM NUCLEO-F767ZI
* STM 32F4DISCOVERY

target boards.

Change-Id: I3c42eb94051352cb3b7eb5b0768a1dc8bdacabce
Reviewed-by: Richard Weickelt <richard@weickelt.de>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>