bare-metal: Long live for IAR ARM toolchain support on Windows
This commit adds a basic support of the IAR toolchain for the
ARM processors on Windows host.

To use it with Qt Creator, it is enougth to add there a desired Kit
with a custom IAR C/C++ compilers, and then set to the Kit's QBS
custom properties page following:

* Key: qbs.toolchain
* Value: iar

Tested with EWARM v8.20.2 on Windows using STM NUCLEO-F767ZI
target board.

Change-Id: I3c42eb94051352cb3b7eb5b0768a1dc8bdacabce