bare-metal: Add IAR 8051 toolchain support
This commit adds a basic support of the IAR Embedded Workbench
toolchain for the 8051 processors family.

To use it with Qt Creator, it is enough to add there a desired Kit
with a custom IAR C/C++ compiler, and then set the following in the
Kit's Qbs profile settings:

* Key: qbs.toolchainType
* Value: iar

Tested with EW for 8051 v10.10.1 on Windows using the simple
projects samples which are comes with the default IAR installer.

Change-Id: I5ef1cc047c27a7fce4a6841331a67e8dd0f0453e
Reviewed-by: Richard Weickelt <richard@weickelt.de>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>