Scanner item: Do not re-use the searchPaths value for other inputs
For the C++ scanner, the search paths depend on module properties, not
on the actual input file. This assumption does not generally hold for
Scanner items, which was not taken into account when this feature was
originally introduced. As a result, the first input to get scanned by a
Scanner item erroneously provided the search paths for all others.

Change-Id: I0159d826dac1f5605494d54cfe96fe468044ca34
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>