bare-metal: Link with library dependencies on IAR for ARM
This patch adds the possibility to link the application with
an external or dependent libraries. IAR supports linking
only with the static libraries for the bare-metal devices.

Now it is possible to use the cpp.staticLibraries and
cpp.libraryPaths properties, also to use the product's
library dependencies via Depends { name: "mylib" }.

Change-Id: I5129473dc0353970d060a8c714b2f4b51d1ab4cc
Reviewed-by: Richard Weickelt <richard@weickelt.de>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>