bare-metal: Add IAR AVR toolchain support
This commit adds a basic support of the IAR Embedded Workbench
toolchain for the AVR processors family.

To use it with Qt Creator, it is enough to add there a desired Kit
with a custom IAR C/C++ compiler, and then set the following in the
Kit's Qbs profile settings:

* Key: qbs.toolchainType
* Value: iar

Tested with EW for AVR v7.20.1 on Windows using the simple
projects samples which are comes with the default IAR installer.

Change-Id: I4ba65e0a753f8703764551c62d630849ff86d91f
Reviewed-by: Richard Weickelt <richard@weickelt.de>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>