bare-metal: Add KEIL ARM toolchain support for Windows
This commit adds a basic support of the KEIL toolchain for the
ARM processors family.

The KEIL toolchain support only the Windows environment:
* http://www.keil.com/support/docs/1456.htm

To use it with Qt Creator, it is enough to add there a desired Kit
with a custom KEIL C/C++ compiler, and then set the following in the
Kit's Qbs profile settings:

* Key: qbs.toolchainType
* Value: keil

Tested with the KEIL uVision v5.23 on Windows using the
STM32 NUCLEO-F767ZI development board.

Change-Id: I93a30f38f9b0e31bf4d1e379a3bdc785e8474ecb
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>