bare-metal: Add SDCC toolchain support
This commit adds a basic support for the SDCC compiler:

 * http://sdcc.sourceforge.net/

As this compiler support multiple architectures, then it is impossible
to uniquely identify the current architecture by dumping of the
pre-defined macros (because its content depends on a target flag). In
this case the cpp.architecture will contains a default architecture
(which is dumped with an omitted target flag).

To use it with Qt Creator, it is enough to add there a desired Kit
with a custom SDCC C/C++ compiler, and then set the following in the
Kit's Qbs profile settings:

* Key: qbs.toolchainType
* Value: sdcc

To create the SDCC profile it is enougth to use the following
command:

 qbs setup-toolchains --type sdcc <path/to/sdcc/compiler/binary> <profile
name>

A toolchain type can be omitted; in this case the QBS will
tries to detect the toolchain type from the specified compiler
name.

Also it is possible to auto-detect the SDCC toolchain from the
PATH environment using the following command:

 qbs setup-toolchain --detect

At current time are supported only the 8051 (aka MCS51) architecture;
other architectures can be added later.

Change-Id: I8cc239d62e35472ab667e054a64a1e59c2d548bd
Reviewed-by: Richard Weickelt <richard@weickelt.de>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>