Fix qbsConfig test on Macos
It seems, that QSettings is not working properly with SystemScope on
Macos (QTBUG-21062). QSettings::fileName returns a path to a file in a
test directory, not in "/Library/Application Support" as expected. Thus,
canWriteToSystemSettings is true and test fails when real qbs binary
tries to write to the /Library... location

Change-Id: I726360a1daacf4871f944f5afe30089083283dce
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>