Add pkgconfig.sysroot property
By default, pkg-config uses qbs.sysroot property which is not desirable
in some cases - for example, on macOS, even build for current system is
sysrooted to the SDK folder within XCode installation. pkg-config
installed with brew resides in /usr/local and can't be found within an
SDK dir. For that reason, it make sense to override it manually

Change-Id: Ibf779fd1bf0e87189abb2eab23d4e9e153fcca17
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>