Enable Travis CI to build and test Qbs
This patch adds a configuration file for building and testing Qbs with Travis
CI. Travis supports Linux, Windows and Mac OS. It requires a repository on
github, but is generally free for open source projects. Users could simply fork
Qbs, connect their Travis account to the forked repository and let it build.

In addition to this patch, a bot has been implemented that watches
codereview.qt-project.org for changes. Whenever a user uploads a new change or
patch set, the bot pushes this change to a mirror repository on github and kicks
Travis to build it. Once the build has finished the result is posted as a
comment back to gerrit.

This bot could be extended to behave like the Qt CI system.

Change-Id: Ic0e1bd507171f40860bfe352651880b4ae7412fc
Reviewed-by: Qbs CI Bot <travis-bot@weickelt.de>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>