Update windowsservercore/Dockerfile
- Update qbs to 1.12.1
- Update Qt to 5.12.5

We're not using the installer anymore, because
a) the unattended-install-hack doesn't work for newer Qt installers
b) newer installers do not provide VS 2015 builds.
c) it allows us to tailor the Qt build to our needs in the future.

We're not building the documentation on Windows anymore due to the
genius move to require libclang for qdoc. That means we need
additional twiddling with the packages to add documentation that was
built on an usable OS.

Task-number: QBS-1438
Change-Id: I77a487562c2f5edd69b8b8a10ef3410ab599dc92
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>