Fix compiling with project.withExamples:true on Mac
First, if qbs.targetPlatform is set, it is desired to use a
corresponding architecture (armv7 or arm64); otherwise x86_64
architecture is used.

Second, when building in Qt Creator, xcode module can't be found (due to
different profile settings) and sysroot is not set to point to the
iPhoneOS.sdk which leads to unability to find some frameworks. Disable
this example when using QtC.

Change-Id: I9fe413e6feb8fe7cfc6ff2e58323b81d8bf06839
Reviewed-by: Qbs CI Bot <travis-bot@weickelt.de>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>