baremetal: Long live the IAR EW project generator for ARM
Right now it is possible to generate a native projects for the
IAR Embedded Workbench IDE, e.g. using the following command:

qbs generate -g iarew8 -d <path/to/build/directory> -f <path/to/qbs/project> profile:<your/qbs/profile>

We need in a valid IAR EW QBS profile, from which the generator take
a desired target architecture and other stuff.

The IAR EW generator has a semi-intelligent logic, it parses a source
QBS project and converts a compiler flags and other stuff to an
appropriate configurations of the native IAR EW project.

Currently it is supported only one 'iarew8' generator which allow to
generate a projects for the IAR EW for ARM for all versions of 8 series.

Tested with the IAR EW for ARM v8.20, v8.32, v8.40, using as the QBS
bare-metal examples, and as other projects.

Change-Id: I2aa24d5a635f1f5a0e645bec8c30053e960d90a3
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>