baremetal: Long live the IAR EW project generator for AVR
Right now it is possible to generate a native projects for the
IAR Embedded Workbench IDE for Microchip AVR architecture,
e.g. using the following command:

qbs generate -g iarew7 -d <path/to/build/directory> -f <path/to/qbs/project> profile:<your/qbs/profile>

We need in a valid IAR EW QBS profile, from which the generator
take a desired target architecture and other stuff.

The IAR EW generator has a semi-intelligent logic, it parses
a source QBS project and converts a compiler flags and other
stuff to an appropriate configurations of the native IAR EW
project.

Currently it is supported only one 'iarew7' generator which
allow to generate a projects for the IAR EW for AVR for all
versions of 7 series.

Tested with the IAR EW for AVR v7.20.1, using as the QBS
bare-metal examples, and as other projects.

Change-Id: Ic5811e579e86b868d5946637db9d8780ef47968b
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>