baremetal: Long live the IAR EW project generator for 8051
Right now it is possible to generate a native projects for the
IAR Embedded Workbench IDE, e.g. using the following command:

qbs generate -g iarew10 -d <path/to/build/directory> -f <path/to/qbs/project> profile:<your/qbs/profile>

We need in a valid IAR EW QBS profile, from which the generator take
a desired target architecture and other stuff.

The IAR EW generator has a semi-intelligent logic, it parses a source
QBS project and converts a compiler flags and other stuff to an
appropriate configurations of the native IAR EW project.

Currently it is supported only one 'iarew10' generator which allow to
generate a projects for the IAR EW for 8051 for all versions of 10 series.

Tested with the IAR EW for 8051 v10.10.1, using as the QBS
bare-metal examples, and as other projects.

Change-Id: If8a7397bebf6176010d42cfda41230e50a499a69
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>