bare-metal: Add IAR STM8 toolchain support
This commit adds a basic support of the IAR Embedded Workbench
toolchain for the STM8 processors family.

To use it with Qt Creator, it is enough to add there a desired Kit
with a custom IAR C/C++ compiler, and then set the following in the
Kit's Qbs profile settings:

* Key: qbs.toolchainType
* Value: iar

Tested with EW for STM8 v3.11.1 on Windows using the simple
projects samples which come with the default IAR installer.

Change-Id: Ibc116ceb33b7df49c3241d26fa97136255ba06dd
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>