bare-metal: Auto-detect SDCC profiles for all supported architectures
Previous SDCC toolchain auto-detection algorithm based on an assumption,
that the default target architecture always is MCS51.

But, the SDCC compiler support multiple architectures:

* http://sdcc.sourceforge.net/

A target architecture sets by the special command-line flag
(f.e. -mmcu51 for 8051, -mstm8 for STM8 and so forth).

As previously we have added the STM8 architecture support to the SDCC
compiler, then it is reasonable to create a multiple SDCC profiles for
one compiler for the different supported architectures.

Right now, the SDCC compiler will be auto-detected as two instances:

* with mcs51 architecture
* with stm8 architecture

Change-Id: I62949a30b770dab005628dc4fbd6ae67f5cf28ff
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>