baremetal: Long live the IAR EW project generator for STM8
Right now it is possible to generate a native projects for the
IAR Embedded Workbench IDE, e.g. using the following command:

qbs generate -g iarew3 -d <path/to/build/directory> -f <path/to/qbs/project> profile:<your/qbs/profile>

We need in a valid IAR EW QBS profile, from which the generator take
a desired target architecture and other stuff.

The IAR EW generator has a semi-intelligent logic, it parses a source
QBS project and converts a compiler flags and other stuff to an
appropriate configurations of the native IAR EW project.

Currently it is supported only one 'iarew3' generator which allow to
generate a projects for the IAR EW for STM8 for all versions of 3 series.

Tested with the IAR EW for STM8 v3.11.1, using as the QBS
bare-metal examples, and as other projects.

Change-Id: I47880d62cd1e81ed7bbfba840a5af0c558065013
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>