baremetal: Long live the KEIL UV project generator for ARM
Right now it is possible to generate a native projects for the
KEIL UVision IDE, e.g. using the following command:

qbs generate -g keiluv5 -d <path/to/build/directory> -f <path/to/qbs/project> profile:<your/qbs/profile>

We need in a valid KEIL QBS profile, from which the generator take
a desired target architecture and other stuff.

The KEIL UV generator has a semi-intelligent logic, it parses a source
QBS project and converts a compiler flags and other stuff to an
appropriate configurations of the native KEIL UVsion project.

Currently it is supported only one 'keiluv5' generator which allow to
generate a projects for the KEIL UVision v5 for ARM architecture.

Tested with the KEIL UVision v5.23 for ARM, using as the QBS
bare-metal examples, and as other projects.

Change-Id: I3af9d01f25a9570a99a62d4ce2c30fec3566b9a1
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>