Prevent from 10min timeout on Travis due to long-running testcases
TestBlackbox runs for more than 10 minutes in the Windows environment on Travis
and since recently also on Mac OS. This might trigger a non-configurable 10
minutes timeout due to no output generated by Qbs. The official workaround is
travis_wait. But it is flawed, because it accumulates all output before it
prints something, thus making it impossible to watch the output of a running
build. This problem is often not noticed because other tests may run in parallel
and produce output.

This patch adds a background process which outputs an empty line every 9:50
minutes while running autotests.

Change-Id: I0ac5d0f581aabbc0d8641a13e7d207dc84902b95
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>