baremetal: Add IAR MSP430 toolchain support
This commit adds a basic support of the IAR Embedded Workbench
toolchain for the MSP430 processors family.

To use it with Qt Creator, it is enough to add there a desired Kit
with a custom IAR C/C++ compiler, and then set the following in the
Kit's Qbs profile settings:

* Key: qbs.toolchainType
* Value: iar

Tested with EW for MSP430 v7.12.4 on Windows using the simple
projects samples which come with the default IAR installer.

Change-Id: I6a31da1823fd08ba7bf41a315ac20e98f5ba77f6
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>