baremetal: Long live the IAR EW project generator for MSP430
Right now it is possible to generate a native projects for the
IAR Embedded Workbench IDE, e.g. using the following command:

qbs generate -g iarew7 -d <path/to/build/directory> -f <path/to/qbs/project> profile:<your/qbs/profile>

We need in a valid IAR EW QBS profile, from which the generator take
a desired target architecture and other stuff.

The IAR EW generator has a semi-intelligent logic, it parses a source
QBS project and converts a compiler flags and other stuff to an
appropriate configurations of the native IAR EW project.

Currently it is supported only one 'iarew7' generator which allow to
generate a projects for the IAR EW for MSP430 for all versions of 7 series.

Tested with the IAR EW for MSP430 v7.12.4, using as the QBS
bare-metal examples, and as other projects.

Change-Id: Ic0bbc82e6ebb3b04bced639797976e432809171f
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>