Clean up version treatment in cpp module
The handling of version numbers in the cpp module was  overly complicated and
too restrictive. Version numbers that are valid according to SemVer (for
instance 1.2.3-beta+47) were not accepted although being valid.

This patch allows the usage of any version number as long as it starts with
x[.y[.z]]. It also simplifies version number handling in the cpp module.

Task-number: QBS-1486
Change-Id: I2790f8a80bd78bf549f5d46c0748f88c788addec
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>