Allow overlapping uid/gid in container
Makes it possible to start the container on a Linux host
where the uid/gid of the current user is similar to an
existing uid/gid in the container that is not the "devel"
user.

For instance, if the host gid is 100, then groupmod in the
entrypoint script refuses to change the "devel" user to gid
100 because that gid is already taken by the "apt" group in
the container.

Change-Id: Ifc4980b118b9b94bf744aa1108225d3c727eb644
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>