baremetal: Don't interpret BL51 linker warnings as errors
BL51 linker (from Keil product for 8051 architecture) ends
with the exit code '1' in case any linking warning occurred.
It is not an error, because at errors the linker ends with
the exit code '2'.

So, we need just to increase the value of the linking command
'maxExitCode' property from 0 to 1, similar to the compiler
command handling.

Change-Id: I4917d08ca3407c8046879780e4c8696fdef7736a
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>