baremetal: Introduce new cpp::generate{Assembler|Compiler}ListingFiles
This properties enables or disables generation of a compiler or
assembler listing files.

Reason why we need in this property is in that some compilers
(e.g. KEIL C51) generates a listing files by default, that spams
a project sources directory. So, we need to have a possibility
e.g. to disable it or to generate to an output directory.

This patch implements this feature for KEIL, IAR, SDCC compilers
with the following restrictions:

* IAR (8051, AVR, STM8, MSP430, ARM) - full support.

* KEIL (8051) - full support.

* KEIL (ARM) - has only one restriction in that a compiler does
  not support specifying of an output listing file name. It is
  possible to specify only an output listing directory. So, a
  listing file names will be a bit different than for other
  compilers (e.g. if a source file name is 'foo.c', then the
  listing file name will be 'foo.lst', instead of 'foo.c.lst').

* SDCC (8051, STM8) - seems, has not possibility to disable an
  auto-generated listing files. But it generates an output listing
  files to a right output directory with the object files. Besides,
  a listing files has a correct names (e.g. for the 'foo.c' file,
  the listing file will be 'foo.c.lst').

Change-Id: Ic3516101e69eed156cf71606a7144efc72d40204
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>