baremetal: Add IAR Renesas RL78 toolchain support
This commit adds a basic support of the IAR Embedded Workbench
toolchain for the Renesas RL78 processors family.

To use it with Qt Creator, it is enough to add there a desired Kit
with a custom IAR C/C++ compiler, and then set the following in the
Kit's Qbs profile settings:

* Key: qbs.toolchainType
* Value: iar

Tested with EW for RL78 v4.10.1 on Windows using the simple
projects samples which come with the default IAR installer.

Change-Id: I1f68f27feaac26a5cf606f326b97f8bccc5c2847
Reviewed-by: Richard Weickelt <richard@weickelt.de>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>