baremetal: Auto-detect MPLAB X32 GCC toolchain on Windows
Microchip Technology Inc. provides set of own GCC toolchains
to use it with the PIC32 microcontrollers:

* https://www.microchip.com/mplab/compilers

On Windows these toolchains distributed in a form of the
installer packages. So, it will be good to auto-detect such
toolchains there.

But, this patch do it partially, it can not detect each
installed toolchain as separate instance.

For example, if we have a two toolchains v2.20 and v2.30
installed, then it will looks like this:

 Profile 'xc32-gcc' created for 'C:/microchip/xc32/v2.20/bin/xc32-gcc.exe'.
 Profile 'xc32-gcc' created for 'C:/microchip/xc32/v2.30/bin/xc32-gcc.exe'.

In resulting, the toolchain v2.20 will be overwritten by
toolchain v2.30, because its names are identical, and then
the one last profile will be created.

Seems, to solve this issue we need to dump the compiler
version. But a problem is in that both compilers has
same 4.8.3 version. So, it is not possible to separate
there two compilers in a simple way. Maybe, in future,
we can look on the __XC_VERSION__ macro, and to interpret
it as 'build' version (which is 2200 vs 2300). But, for
this purpose we need to refactor the logic in gccprobe.cpp.

Change-Id: Id804a411ed6046b31e57b0468d75830eef92693c
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>