Kill the keep-alive background process
The sleep command in background was not killed although its parent process was.
That caused builds on Travis to idle for up to 590s after completion in
worst-case.

This patch extends the EXIT trap and ensures that the sleep command is killed.
There is unfortunately no generic way in Bash to achieve that (without writing
lots of convoluted code) because utilities like pkill are not available in Git
Bash.

Change-Id: Ibbb5683e57c242e58133c1860ec9c921b7e449b7
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>