MSVC: Remember only dependencies with libraries
... when collecting link-time dependencies.
This is a hack for the case where an application depends on a static
library, and both depend on a statically built Qt. A Qt plugin does not
declare a library artifact when pulled in by a static library, because
we don't want to link against plugins in that case. As a result, when
the dependency collection runs for the application and finds a Qt
plugin via traversing the static library, it won't see any libraries
there. Before this patch, we'd skip such plugins when encountering
them again as direct dependencies of the application, so they were
missing on the linker command line.

Fixes: QBS-1518
Change-Id: Ia0e5b39e4a2fb80b9fd1384610942ac627978571
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>