API: Do not try to do manual book-keeping
... when adding/removing files to/from a project.
We originally implemented this so that IDE users could edit their
project files via the UI without having to re-resolve the project, thus
saving time.
However, great care has to be taken to ensure that we really do the same
thing as would happen on a proper resolve step, and we do not seem to
manage, as demonstrated by the crashes I have observed recently when
removing files via Qt Creator.
So let's play it safe and remove the shortcut.

Change-Id: I38e7dcdba0aa129f43feec084ae140e1a57d2ee9
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>