doc: Provide a bit more info in the README file
Qbs has a mirror on the GitHub.com and the README file is a starting
point for people who come to the website. It is nice to have quick
links that can be useful for newcomers.

Also, rename the README file to README.md to be able to use Markdown

Change-Id: I38394cf0a0802bfebebc9f070732041cfe24090c
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>