Add Docker environment to test Qbs with Qt for Android
The Docker images contains ndk r20 and qt 5.13.2 as well as 5.14.0. Containers
can be started with the following command line:

  - docker-compose pull bionic-android-513
  - docker-compose run --rm bionic-android-513

The current folder is mounted under /qbs. Assuming that Qbs has been built in
the current directory and with release configuration, the autotests can be
invoked with:

    docker-compose run --rm bionic-android-513 scripts/test-qt-for-android.sh \
        release/install-root/usr/local/bin

A job Travis CI job is added which builds Qt and runs the Android blackbox
tests.

Change-Id: I76b1901cb7c41f8b5b0122265c99c9b37b59f261
Reviewed-by: Richard Weickelt <richard@weickelt.de>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>