Do not override QT_INSTALL_DIR in the Travis Windows job
QT_INSTALL_DIR is defined as global environment variable and is overrideen with
a different value in the windows job. Something changed in the way Travis parses
environment variables and the override is exported after depend variables.

Since we run the build in a bash shell, there is no reason to set QT_INSTALL_DIR
to a different value than ~/Qt. We can safely remove the override.

Change-Id: Id556997b3a9d8a8d6be474485b5084955a09f253
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>