Use docker-compose exclusively for managing Docker images
Qbs used to have docker products and a module for building the images, but this
became obsolet and was superseded by the docker-compose infrastructure.
Docker-compose is a much better fit for our use-case (which is mostly CI). It
allows us to use unique (and complicated) tags both when building and running
the images without exposing that to the user. Unique tags like
bionic-5.12.6_1.14.1-2 instead of just bionic ensure that changes in the
infrastructure don't suddenly break builds in older branches using the same
image.

This patch also updates the documentation to emphasize the usage of
docker-compose.

Change-Id: I45335d37034d10dda5ab5079ee3be1f81cb3e5c5
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>