Export framework search path in qtscxml blackbox test on macOS
The qtscxml testcase builds a Qt binary and runs it. This relies on library
search paths being correctly exported so that the Qt libraries and framework
files are found by the binary. The testcase exports LD_LIBRARY_PATH, but does
neither export DYLD_LIBRARY_PATH nor DYLD_FRAMEWORK_PATH which is needed on
macOS as well. Therefore the binary cannot be loaded.

This error is only visible when running tst_blackbox-qt stand-alone. It is
usually hidden because the QbsAutoTestRunner exports both DYLD variables. But if
the user builds the autotests against Qt installation A while testing
installation B, then the binary in the qtscxml testcase uses the wrong
DYLD_LIBRARY_PATH.

This patch exports both DYLD_LIBRARY_PATH and DYLD_FRAMEWORK_PATH on macOS.

Change-Id: I348d50ec0417e35ff379aad2f93cf80b5ed50c95
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>