Refactor MSBuild lib
This patch modernizes the msbuild lib by replacing
raw news with parent-child ownership with unique_ptr.

The reason is that tools like clang analyzer do not
understand the parent-child trees and  provide false
positives in the code.

This patch also introduces a check for an allowlist
of children types avalibble for a node. Previously,
this check was done implicitly via the type of parent
node.

Change-Id: Id5f589271711be1ca51ce42098132d20cd274cd4
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>