Exclude VariantValues when checking for conflicts
Qbs prints a warning when different modules contain conflicting assignments to
scalar properties. This check assumed that the source values were always of type
JSSourceValue and thus performed a invalid static cast which caused sporadic
crashes on macOS in TestLanguage::moduleMergingVariantValues. No crashes were
obtained on other platforms, but that was only a coincidence.

This patch amends 8ff1dd00.

Change-Id: Id4809496a5fbc4985d040118a4684f3ac040cdc0
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>