baremetal: Fix handling cpp.libraryPaths for IAR toolchain
An IAR toolchain contains two different linkers (xlinker and ilinker),
which are depend on a target architecture. These linkers have different
command line options, one of which is a libraries search paths:

* '-L' for xlinker
* '-I' for ilinker

But we always used '-L' flag, that was wrong, because we need to separate
it only for xlinker.

Change-Id: Ib2b482d4e9ac4aaefdf47c81f16c6c389243f56d
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>