baremetal: Simplify output artifact extensions handling for IAR toolchain
Some architectures in IAR have 'magic' code names, e.g. "avr" equals
with '90', "avr32" equals with '82' and so forth. Besides, an output
artifacts for executable/object/library objects contains that 'magic'
codes in own file extensions.

So, it is makes sense to create the separate function which return
that 'magic' codes and to use this function inside of
executable/object/library extensions generation to simplify a code
and to minimize the copy/paste operations.

Change-Id: Id2223bc577847c616f58bc4eb736d3192a3285a7
Reviewed-by: Ivan Komissarov <ABBAPOH@gmail.com>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>