XCode: fix handling provisioning profiles
The abuse of the file tags to pass additional information to the prepare
script is not working anymore, so set a fake module property instead.

Also, automatically tag all provisioning profiles located in
xcode.provisioningProfilesPath.

Also, do not raise error if user did not provide any *.entitlements
files.

This patch fixes deploying the iOS application to the device

Change-Id: I9356af334a928099d59f747d6f186634df8168d3
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>