Msvc: Handle windows module definition file
Module-definition (.def) file provide the linker with information about
exports, attributes, and other information about the program to be linked.

This file can be used instead of e.g. __declspec(dllexport) keywords
inside of a header files to export an appropriate symbols:

 * https://docs.microsoft.com/en-us/cpp/build/reference/module-definition-dot-def-files?view=vs-2019
 * https://docs.microsoft.com/en-us/cpp/build/reference/def-specify-module-definition-file?view=vs-2019

To make it supported, we have created the additional 'def' file tag,
and now we can add the required module-definition file to the
{Product|Group}'s 'files' property.

Fixes: QBS-571
Change-Id: I831d2e1b8f9f061a4414a5eaac85b71584f48c5d
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>